fix(backends): correct emitter output across java, go, zig, php and ruby
java: serialize Duration as the real wire shape via paired Jackson
converters, on both the record component and the builder setter, since
@JsonDeserialize(builder) bypasses the record's own annotations.
go: derive snippet call shape from the extracted FunctionDef instead of
re-asserting it in config, and infer ptr(N) from the destination field.
zig: wrap an optional capsule return, use the correctly prefixed visitor
callbacks struct, and back VisitorHandle with u64 rather than *anyopaque.
php: sort constructor params required-before-optional to match the stub.
php/extendr: stop suppressing an enum variant factory on a name collision
with enum_def.methods -- no backend forwards those, so the factory was
dropped with nothing replacing it.
magnus/php/napi: stop declaring in stubs what the binding generator does
not emit; the stub side now consults the binding side's own predicate.

+++ src/backends/php/gen_bindings/php_types.rs
@@ -76,7 +76,9 @@ pub(super) fn php_type(ty: &TypeRef) -> String {
         TypeRef::Vec(_) | TypeRef::Map(_, _) => "array".to_string(),
         TypeRef::Named(name) => name.clone(),
         TypeRef::Unit => "void".to_string(),
-        TypeRef::Duration => "float".to_string(),
+        // Duration crosses the FFI boundary as milliseconds (Rust `i64`, see `PhpMapper::duration`
+        // in `type_map.rs`), not seconds — the stub type is `int`, not `float`. ~keep
+        TypeRef::Duration => "int".to_string(),
     }
 }
